#14 Move pages to seperate modules
